% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{Encode DNA Sequences into Feature Matrix}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector of DNA 5-mer strings.}
}
\value{
A data frame with nucleotide positions (nt_pos1...nt_posN) as columns and factor levels (A, T, C, G).
}
\description{
This function converts DNA 5-mer sequences into a position-wise categorical feature matrix.
}
\examples{
dna_encoding(c("ATCGA", "TGCAT"))
}
