% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict m6A Sites for Multiple Samples}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained random forest model (e.g. rf_fit.rds).}

\item{feature_df}{A data frame containing features including gc_content, RNA_type, RNA_region,
exon_length, distance_to_junction, evolutionary_conservation, and DNA_5mer.}

\item{positive_threshold}{Numeric. Threshold above which samples are labeled as "Positive" (default = 0.5).}
}
\value{
A data frame with predicted probabilities and labels added.
}
\description{
Predict m6A modification probability and status for multiple samples using a trained ML model.
}
\examples{
\dontrun{
mdl <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
df  <- read.csv(system.file("extdata", "m6A_input_example.csv", package="m6APrediction"))
prediction_multiple(mdl, df, positive_threshold = 0.6)
}
}
