% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict m6A Sites for a Single Sample}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A trained random forest model (e.g. rf_fit.rds).}

\item{gc_content}{GC content (numeric).}

\item{RNA_type}{RNA type (character: "mRNA", "lincRNA", "lncRNA", or "pseudogene").}

\item{RNA_region}{RNA region (character: "CDS", "intron", "3'UTR", or "5'UTR").}

\item{exon_length}{Exon length (numeric).}

\item{distance_to_junction}{Distance to exon–intron junction (numeric).}

\item{evolutionary_conservation}{Evolutionary conservation score (numeric).}

\item{DNA_5mer}{DNA 5-mer sequence (character).}

\item{positive_threshold}{Threshold for classifying "Positive"/"Negative" (default = 0.5).}
}
\value{
A named vector with predicted probability and status.
}
\description{
This function predicts m6A probability and classification for a single RNA feature input.
}
\examples{
\dontrun{
mdl <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
prediction_single(
  mdl,
  gc_content = 0.6,
  RNA_type = "mRNA",
  RNA_region = "CDS",
  exon_length = 12,
  distance_to_junction = 5,
  evolutionary_conservation = 0.8,
  DNA_5mer = "ATCGT",
  positive_threshold = 0.5
)
}
}
